Module aws_sdk_mturk::client::fluent_builders

Utilities to ergonomically construct a request to the service.

Fluent builders are created through the Client by calling one if its operation methods. After parameters are set using the builder methods, the send method can be called to initiate the request.

Structs

AcceptQualificationRequest

Fluent builder constructing a request to AcceptQualificationRequest.

ApproveAssignment

Fluent builder constructing a request to ApproveAssignment.

AssociateQualificationWithWorker

Fluent builder constructing a request to AssociateQualificationWithWorker.

CreateAdditionalAssignmentsForHIT

Fluent builder constructing a request to CreateAdditionalAssignmentsForHIT.

CreateHIT

Fluent builder constructing a request to CreateHIT.

CreateHITType

Fluent builder constructing a request to CreateHITType.

CreateHITWithHITType

Fluent builder constructing a request to CreateHITWithHITType.

CreateQualificationType

Fluent builder constructing a request to CreateQualificationType.

CreateWorkerBlock

Fluent builder constructing a request to CreateWorkerBlock.

DeleteHIT

Fluent builder constructing a request to DeleteHIT.

DeleteQualificationType

Fluent builder constructing a request to DeleteQualificationType.

DeleteWorkerBlock

Fluent builder constructing a request to DeleteWorkerBlock.

DisassociateQualificationFromWorker

Fluent builder constructing a request to DisassociateQualificationFromWorker.

GetAccountBalance

Fluent builder constructing a request to GetAccountBalance.

GetAssignment

Fluent builder constructing a request to GetAssignment.

GetFileUploadURL

Fluent builder constructing a request to GetFileUploadURL.

GetHIT

Fluent builder constructing a request to GetHIT.

GetQualificationScore

Fluent builder constructing a request to GetQualificationScore.

GetQualificationType

Fluent builder constructing a request to GetQualificationType.

ListAssignmentsForHIT

Fluent builder constructing a request to ListAssignmentsForHIT.

ListBonusPayments

Fluent builder constructing a request to ListBonusPayments.

ListHITs

Fluent builder constructing a request to ListHITs.

ListHITsForQualificationType

Fluent builder constructing a request to ListHITsForQualificationType.

ListQualificationRequests

Fluent builder constructing a request to ListQualificationRequests.

ListQualificationTypes

Fluent builder constructing a request to ListQualificationTypes.

ListReviewPolicyResultsForHIT

Fluent builder constructing a request to ListReviewPolicyResultsForHIT.

ListReviewableHITs

Fluent builder constructing a request to ListReviewableHITs.

ListWorkerBlocks

Fluent builder constructing a request to ListWorkerBlocks.

ListWorkersWithQualificationType

Fluent builder constructing a request to ListWorkersWithQualificationType.

NotifyWorkers

Fluent builder constructing a request to NotifyWorkers.

RejectAssignment

Fluent builder constructing a request to RejectAssignment.

RejectQualificationRequest

Fluent builder constructing a request to RejectQualificationRequest.

SendBonus

Fluent builder constructing a request to SendBonus.

SendTestEventNotification

Fluent builder constructing a request to SendTestEventNotification.

UpdateExpirationForHIT

Fluent builder constructing a request to UpdateExpirationForHIT.

UpdateHITReviewStatus

Fluent builder constructing a request to UpdateHITReviewStatus.

UpdateHITTypeOfHIT

Fluent builder constructing a request to UpdateHITTypeOfHIT.

UpdateNotificationSettings

Fluent builder constructing a request to UpdateNotificationSettings.

UpdateQualificationType

Fluent builder constructing a request to UpdateQualificationType.