Struct aws_sdk_mturk::client::fluent_builders::UpdateQualificationType

pub struct UpdateQualificationType<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

Fluent builder constructing a request to UpdateQualificationType.

The UpdateQualificationType operation modifies the attributes of an existing Qualification type, which is represented by a QualificationType data structure. Only the owner of a Qualification type can modify its attributes.

Most attributes of a Qualification type can be changed after the type has been created. However, the Name and Keywords fields cannot be modified. The RetryDelayInSeconds parameter can be modified or added to change the delay or to enable retries, but RetryDelayInSeconds cannot be used to disable retries.

You can use this operation to update the test for a Qualification type. The test is updated based on the values specified for the Test, TestDurationInSeconds and AnswerKey parameters. All three parameters specify the updated test. If you are updating the test for a type, you must specify the Test and TestDurationInSeconds parameters. The AnswerKey parameter is optional; omitting it specifies that the updated test does not have an answer key.

If you omit the Test parameter, the test for the Qualification type is unchanged. There is no way to remove a test from a Qualification type that has one. If the type already has a test, you cannot update it to be AutoGranted. If the Qualification type does not have a test and one is provided by an update, the type will henceforth have a test.

If you want to update the test duration or answer key for an existing test without changing the questions, you must specify a Test parameter with the original questions, along with the updated values.

If you provide an updated Test but no AnswerKey, the new test will not have an answer key. Requests for such Qualifications must be granted manually.

You can also update the AutoGranted and AutoGrantedValue attributes of the Qualification type.

Implementations

impl<C, M, R> UpdateQualificationType<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub async fn send(
self
) -> Result<UpdateQualificationTypeOutput, SdkError<UpdateQualificationTypeError>> where
R::Policy: SmithyRetryPolicy<UpdateQualificationTypeInputOperationOutputAlias, UpdateQualificationTypeOutput, UpdateQualificationTypeError, UpdateQualificationTypeInputOperationRetryAlias>,

Sends the request and returns the response.

If an error occurs, an SdkError will be returned with additional details that can be matched against.

By default, any retryable failures will be retried twice. Retry behavior is configurable with the RetryConfig, which can be set when configuring the client.

pub fn qualification_type_id(self, inp: impl Into<String>) -> Self

The ID of the Qualification type to update.

pub fn set_qualification_type_id(self, input: Option<String>) -> Self

The ID of the Qualification type to update.

pub fn description(self, inp: impl Into<String>) -> Self

The new description of the Qualification type.

pub fn set_description(self, input: Option<String>) -> Self

The new description of the Qualification type.

pub fn qualification_type_status(self, inp: QualificationTypeStatus) -> Self

The new status of the Qualification type - Active | Inactive

pub fn set_qualification_type_status(
self,
input: Option<QualificationTypeStatus>
) -> Self

The new status of the Qualification type - Active | Inactive

pub fn test(self, inp: impl Into<String>) -> Self

The questions for the Qualification test a Worker must answer correctly to obtain a Qualification of this type. If this parameter is specified, TestDurationInSeconds must also be specified.

Constraints: Must not be longer than 65535 bytes. Must be a QuestionForm data structure. This parameter cannot be specified if AutoGranted is true.

Constraints: None. If not specified, the Worker may request the Qualification without answering any questions.

pub fn set_test(self, input: Option<String>) -> Self

The questions for the Qualification test a Worker must answer correctly to obtain a Qualification of this type. If this parameter is specified, TestDurationInSeconds must also be specified.

Constraints: Must not be longer than 65535 bytes. Must be a QuestionForm data structure. This parameter cannot be specified if AutoGranted is true.

Constraints: None. If not specified, the Worker may request the Qualification without answering any questions.

pub fn answer_key(self, inp: impl Into<String>) -> Self

The answers to the Qualification test specified in the Test parameter, in the form of an AnswerKey data structure.

pub fn set_answer_key(self, input: Option<String>) -> Self

The answers to the Qualification test specified in the Test parameter, in the form of an AnswerKey data structure.

pub fn test_duration_in_seconds(self, inp: i64) -> Self

The number of seconds the Worker has to complete the Qualification test, starting from the time the Worker requests the Qualification.

pub fn set_test_duration_in_seconds(self, input: Option<i64>) -> Self

The number of seconds the Worker has to complete the Qualification test, starting from the time the Worker requests the Qualification.

pub fn retry_delay_in_seconds(self, inp: i64) -> Self

The amount of time, in seconds, that Workers must wait after requesting a Qualification of the specified Qualification type before they can retry the Qualification request. It is not possible to disable retries for a Qualification type after it has been created with retries enabled. If you want to disable retries, you must dispose of the existing retry-enabled Qualification type using DisposeQualificationType and then create a new Qualification type with retries disabled using CreateQualificationType.

pub fn set_retry_delay_in_seconds(self, input: Option<i64>) -> Self

The amount of time, in seconds, that Workers must wait after requesting a Qualification of the specified Qualification type before they can retry the Qualification request. It is not possible to disable retries for a Qualification type after it has been created with retries enabled. If you want to disable retries, you must dispose of the existing retry-enabled Qualification type using DisposeQualificationType and then create a new Qualification type with retries disabled using CreateQualificationType.

pub fn auto_granted(self, inp: bool) -> Self

Specifies whether requests for the Qualification type are granted immediately, without prompting the Worker with a Qualification test.

Constraints: If the Test parameter is specified, this parameter cannot be true.

pub fn set_auto_granted(self, input: Option<bool>) -> Self

Specifies whether requests for the Qualification type are granted immediately, without prompting the Worker with a Qualification test.

Constraints: If the Test parameter is specified, this parameter cannot be true.

pub fn auto_granted_value(self, inp: i32) -> Self

The Qualification value to use for automatically granted Qualifications. This parameter is used only if the AutoGranted parameter is true.

pub fn set_auto_granted_value(self, input: Option<i32>) -> Self

The Qualification value to use for automatically granted Qualifications. This parameter is used only if the AutoGranted parameter is true.