Struct aws_sdk_mturk::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for MTurkRequesterServiceV20170117.

This client allows ergonomic access to a MTurkRequesterServiceV20170117-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

pub fn conf(&self) -> &Config

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn accept_qualification_request(
&self
) -> AcceptQualificationRequest<C, M, R>

pub fn approve_assignment(&self) -> ApproveAssignment<C, M, R>

pub fn associate_qualification_with_worker(
&self
) -> AssociateQualificationWithWorker<C, M, R>

pub fn create_additional_assignments_for_hit(
&self
) -> CreateAdditionalAssignmentsForHIT<C, M, R>

pub fn create_hit(&self) -> CreateHIT<C, M, R>

pub fn create_hit_type(&self) -> CreateHITType<C, M, R>

pub fn create_hit_with_hit_type(&self) -> CreateHITWithHITType<C, M, R>

pub fn create_qualification_type(&self) -> CreateQualificationType<C, M, R>

pub fn create_worker_block(&self) -> CreateWorkerBlock<C, M, R>

pub fn delete_hit(&self) -> DeleteHIT<C, M, R>

pub fn delete_qualification_type(&self) -> DeleteQualificationType<C, M, R>

pub fn delete_worker_block(&self) -> DeleteWorkerBlock<C, M, R>

pub fn disassociate_qualification_from_worker(
&self
) -> DisassociateQualificationFromWorker<C, M, R>

pub fn get_account_balance(&self) -> GetAccountBalance<C, M, R>

pub fn get_assignment(&self) -> GetAssignment<C, M, R>

pub fn get_file_upload_url(&self) -> GetFileUploadURL<C, M, R>

pub fn get_hit(&self) -> GetHIT<C, M, R>

pub fn get_qualification_score(&self) -> GetQualificationScore<C, M, R>

pub fn get_qualification_type(&self) -> GetQualificationType<C, M, R>

pub fn list_assignments_for_hit(&self) -> ListAssignmentsForHIT<C, M, R>

pub fn list_bonus_payments(&self) -> ListBonusPayments<C, M, R>

pub fn list_hi_ts(&self) -> ListHITs<C, M, R>

pub fn list_hi_ts_for_qualification_type(
&self
) -> ListHITsForQualificationType<C, M, R>

pub fn list_qualification_requests(&self) -> ListQualificationRequests<C, M, R>

pub fn list_qualification_types(&self) -> ListQualificationTypes<C, M, R>

pub fn list_reviewable_hi_ts(&self) -> ListReviewableHITs<C, M, R>

pub fn list_review_policy_results_for_hit(
&self
) -> ListReviewPolicyResultsForHIT<C, M, R>

pub fn list_worker_blocks(&self) -> ListWorkerBlocks<C, M, R>

pub fn list_workers_with_qualification_type(
&self
) -> ListWorkersWithQualificationType<C, M, R>

pub fn notify_workers(&self) -> NotifyWorkers<C, M, R>

pub fn reject_assignment(&self) -> RejectAssignment<C, M, R>

pub fn reject_qualification_request(
&self
) -> RejectQualificationRequest<C, M, R>

pub fn send_bonus(&self) -> SendBonus<C, M, R>

pub fn send_test_event_notification(&self) -> SendTestEventNotification<C, M, R>

pub fn update_expiration_for_hit(&self) -> UpdateExpirationForHIT<C, M, R>

pub fn update_hit_review_status(&self) -> UpdateHITReviewStatus<C, M, R>

pub fn update_hit_type_of_hit(&self) -> UpdateHITTypeOfHIT<C, M, R>

pub fn update_notification_settings(
&self
) -> UpdateNotificationSettings<C, M, R>

pub fn update_qualification_type(&self) -> UpdateQualificationType<C, M, R>

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

pub fn from_conf(conf: Config) -> Self