Module aws_sdk_mturk::model

Modules

assignment

See Assignment

bonus_payment

See BonusPayment

hit

See Hit

hit_layout_parameter

See HitLayoutParameter

locale

See Locale

notification_specification

See NotificationSpecification

notify_workers_failure_status

See NotifyWorkersFailureStatus

parameter_map_entry

See ParameterMapEntry

policy_parameter

See PolicyParameter

qualification

See Qualification

qualification_request

See QualificationRequest

qualification_requirement

See QualificationRequirement

qualification_type

See QualificationType

review_action_detail

See ReviewActionDetail

review_policy

See ReviewPolicy

review_report

See ReviewReport

review_result_detail

See ReviewResultDetail

worker_block

See WorkerBlock

Structs

Assignment

The Assignment data structure represents a single assignment

BonusPayment

An object representing a Bonus payment paid to a Worker.

Hit

The HIT data structure represents a single HIT, including

HitLayoutParameter

The HITLayoutParameter data structure defines parameter

Locale

The Locale data structure represents a geographical region or location.

NotificationSpecification

The NotificationSpecification data structure describes a HIT

NotifyWorkersFailureStatus

When MTurk encounters an issue with notifying the Workers

ParameterMapEntry

This data structure is the data type for the AnswerKey

PolicyParameter

Name of the parameter from the Review policy.

Qualification

The Qualification data structure represents a Qualification

QualificationRequest

The QualificationRequest data structure represents a request

QualificationRequirement

QualificationType

The QualificationType data structure represents a

ReviewActionDetail

Both the AssignmentReviewReport and the HITReviewReport

ReviewPolicy

HIT Review Policy data structures represent HIT review

ReviewReport

Contains both ReviewResult and ReviewAction elements for a

ReviewResultDetail

This data structure is returned multiple times for each

WorkerBlock

The WorkerBlock data structure represents a Worker who has

Enums

AssignmentStatus

Comparator

EventType

HitAccessActions

HitReviewStatus

HitStatus

NotificationTransport

NotifyWorkersFailureCode

QualificationStatus

QualificationTypeStatus

ReviewActionStatus

ReviewPolicyLevel

ReviewableHitStatus