Module aws_sdk_mturk::operation

Structs

AcceptQualificationRequest

ApproveAssignment

AssociateQualificationWithWorker

CreateAdditionalAssignmentsForHIT

CreateHIT

The CreateHIT operation creates a new Human Intelligence Task (HIT).

CreateHITType

CreateHITWithHITType

CreateQualificationType

CreateWorkerBlock

The CreateWorkerBlock operation allows you to prevent a Worker from working on your HITs. For example, you can block a Worker who is producing poor quality work. You can block up to 100,000 Workers.

DeleteHIT

DeleteQualificationType

DeleteWorkerBlock

The DeleteWorkerBlock operation allows you to reinstate a blocked Worker to work on your HITs. This operation reverses the effects of the CreateWorkerBlock operation. You need the Worker ID to use this operation. If the Worker ID is missing or invalid, this operation fails and returns the message “WorkerId is invalid.” If the specified Worker is not blocked, this operation returns successfully.

DisassociateQualificationFromWorker

GetAccountBalance

The GetAccountBalance operation retrieves the Prepaid HITs balance in your Amazon Mechanical Turk account if you are a Prepaid Requester.

GetAssignment

GetFileUploadURL

GetHIT

GetQualificationScore

GetQualificationType

ListAssignmentsForHIT

ListBonusPayments

ListHITs

ListHITsForQualificationType

ListQualificationRequests

ListQualificationTypes

ListReviewPolicyResultsForHIT

ListReviewableHITs

ListWorkerBlocks

The ListWorkersBlocks operation retrieves a list of Workers who are blocked from working on your HITs.

ListWorkersWithQualificationType

NotifyWorkers

RejectAssignment

RejectQualificationRequest

SendBonus

SendTestEventNotification

UpdateExpirationForHIT

UpdateHITReviewStatus

UpdateHITTypeOfHIT

UpdateNotificationSettings

UpdateQualificationType